myproject/requirements.txt
# Python dependencies needed for the project, ensuring replicability across different setups

aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Django==4.2.3
django-cors-headers==4.2.0
django-storages==1.13.2
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
joblib==1.3.1
multidict==6.0.4
nltk==3.8.1
openai==0.27.8
PyPDF2==3.0.1
pytz==2023.3
regex==2023.6.3
requests==2.31.0
sqlparse==0.4.4
tqdm==4.65.0
urllib3==1.26.6
yarl==1.9.2
